Add StagedBuffers for device local VkBuffer

+++ src/main/java/no/stelar7/vulkan/engine/spec/UniformSpec.java
@@ -39,7 +39,7 @@ public final class UniformSpec
     private static final Matrix4f mvc = null;
     // UBO Elements end
     
-    public static long getSizeInBytes()
+    public static int getSizeInBytes()
     {
         int size = 0;
